2.ChatModels/chatmodel_anthropic.py
from langchain_anthropic import ChatAnthropic
from dotenv import load_dotenv

load_dotenv()

model = ChatAnthropic(model = 'claude-2')
result = model.invoke('what is the capital of delhi')

print(result.content)