requirements.txt
# Langchain Core
langchain
langchain-core

# OpenAI integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic
langchain-google-genai

# Google Gemeni (PaLM) integration
google-generativeai
langchain-huggingface

# huggingface Face Integration
transformers
huggingface-hub

# Environment management
python-dotenv

# Machine Learning utilities
numpy
scikit-learn
